requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Flask-Script==0.5.1
Flask-Admin==1.0.4
Flask-Login==0.1.3
Flask-WTF==0.8.2
celery==3.0.12
celery-with-redis==3.0
lxml==3.0.2
python-amazon-product-api==0.2.5
python-oembed==0.2.1
redis==2.7.2
